requirements.txt
discord.py==1.6.0
googlesearch-python==2020.0.2
mysql-connector-python==8.0.23